+++ tests/rendered_output.rs
@@ -1,14 +1,18 @@
 extern crate mdbook;
 extern crate tempdir;
 
+mod dummy;
 mod helpers;
+
+use dummy::DummyBook;
+use helpers::assert_contains_strings;
 use mdbook::MDBook;
 
 
 /// Make sure you can load the dummy book and build it without panicking.
 #[test]
 fn build_the_dummy_book() {
-    let temp = helpers::DummyBook::default().build();
+    let temp = DummyBook::default().build();
     let mut md = MDBook::new(temp.path());
 
     md.build().unwrap();
@@ -16,7 +20,7 @@ fn build_the_dummy_book() {
 
 #[test]
 fn by_default_mdbook_generates_rendered_content_in_the_book_directory() {
-    let temp = helpers::DummyBook::default().build();
+    let temp = DummyBook::default().build();
     let mut md = MDBook::new(temp.path());
 
     assert!(!temp.path().join("book").exists());
@@ -28,7 +32,7 @@ fn by_default_mdbook_generates_rendered_content_in_the_book_directory() {
 
 #[test]
 fn make_sure_bottom_level_files_contain_links_to_chapters() {
-    let temp = helpers::DummyBook::default().build();
+    let temp = DummyBook::default().build();
     let mut md = MDBook::new(temp.path());
     md.build().unwrap();
 
@@ -44,13 +48,13 @@ fn make_sure_bottom_level_files_contain_links_to_chapters() {
     let files_in_bottom_dir = vec!["index.html", "intro.html", "second.html", "conclusion.html"];
 
     for filename in files_in_bottom_dir {
-        helpers::assert_contains_strings(dest.join(filename), &links);
+        assert_contains_strings(dest.join(filename), &links);
     }
 }
 
 #[test]
 fn check_correct_cross_links_in_nested_dir() {
-    let temp = helpers::DummyBook::default().build();
+    let temp = DummyBook::default().build();
     let mut md = MDBook::new(temp.path());
     md.build().unwrap();
 
@@ -67,23 +71,23 @@ fn check_correct_cross_links_in_nested_dir() {
     let files_in_nested_dir = vec!["index.html", "nested.html"];
 
     for filename in files_in_nested_dir {
-        helpers::assert_contains_strings(first.join(filename), &links);
+        assert_contains_strings(first.join(filename), &links);
     }
 }
 
 #[test]
 fn rendered_code_has_playpen_stuff() {
-    let temp = helpers::DummyBook::default().build();
+    let temp = DummyBook::default().build();
     let mut md = MDBook::new(temp.path());
     md.build().unwrap();
 
     let nested = temp.path().join("book/first/nested.html");
     let playpen_class = vec![r#"class="playpen""#];
 
-    helpers::assert_contains_strings(nested, &playpen_class);
+    assert_contains_strings(nested, &playpen_class);
 
     let book_js = temp.path().join("book/book.js");
-    helpers::assert_contains_strings(book_js, &[".playpen"]);
+    assert_contains_strings(book_js, &[".playpen"]);
 }
 
 #[test]
@@ -96,7 +100,7 @@ fn chapter_content_appears_in_rendered_document() {
         ("conclusion.html", "Conclusion"),
     ];
 
-    let temp = helpers::DummyBook::default().build();
+    let temp = DummyBook::default().build();
     let mut md = MDBook::new(temp.path());
     md.build().unwrap();
 
@@ -104,6 +108,6 @@ fn chapter_content_appears_in_rendered_document() {
 
     for (filename, text) in content {
         let path = destination.join(filename);
-        helpers::assert_contains_strings(path, &[text]);
+        assert_contains_strings(path, &[text]);
     }
 }
