+++ tests/testing.rs
@@ -1,13 +1,15 @@
-extern crate tempdir;
 extern crate mdbook;
+extern crate tempdir;
 
-mod helpers;
+mod dummy;
+
+use dummy::DummyBook;
 use mdbook::MDBook;
 
 
 #[test]
 fn mdbook_can_correctly_test_a_passing_book() {
-    let temp = helpers::DummyBook::default()
+    let temp = DummyBook::default()
         .with_passing_test(true)
         .build();
     let mut md = MDBook::new(temp.path());
@@ -17,7 +19,7 @@ fn mdbook_can_correctly_test_a_passing_book() {
 
 #[test]
 fn mdbook_detects_book_with_failing_tests() {
-    let temp = helpers::DummyBook::default()
+    let temp = DummyBook::default()
         .with_passing_test(false)
         .build();
     let mut md: MDBook = MDBook::new(temp.path());
